+++ gradle.properties
@@ -2,3 +2,4 @@
 # This is required to provide enough memory for the Minecraft decompilation process.
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
+mod_version=0.0.1
