+++ src/main/java/com/nitro/deeperdarker/registry/DDSounds.java
@@ -1,6 +1,6 @@
-package com.nitro.deeperdarker.core.registry;
+package com.nitro.deeperdarker.registry;
 
-import com.nitro.deeperdarker.core.DeeperAndDarker;
+import com.nitro.deeperdarker.DeeperAndDarker;
 import com.teamabnormals.blueprint.core.util.registry.SoundSubRegistryHelper;
 
 public class DDSounds {
